replace Counter with attestor
