+++ .env
@@ -3,4 +3,4 @@ ANVIL_FORK_URL=https://cloudflare-eth.com
 ETHERSCAN_API_KEY=
 FORGE_RPC_URL=
 FORGE_PRIVATE_KEY=
-VITE_ALCHEMY_API_KEY=ALCHEMY_KEY_HERE
+VITE_ALCHEMY_API_KEY=GaQo9-awAuExHrM9ZXd3QqR4FPydpyWX
