+++ contracts/script/AttestationStation.s.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.13;
 
 import "forge-std/Script.sol";
 
-contract CounterScript is Script {
+contract AttestationStationScript is Script {
     function setUp() public {}
 
     function run() public {
