--- contracts/src/Counter.sol
@@ -1,16 +1,0 @@
-// SPDX-License-Identifier: UNLICENSED
-pragma solidity ^0.8.15;
-
-contract Counter {
-    uint256 public number = 0;
-
-    event Transfer(address indexed from, address indexed to, uint256 value);
-
-    function setNumber(uint256 newNumber) public {
-        number = newNumber;
-    }
-
-    function increment() public {
-        number++;
-    }
-}
