+++ src/generated.ts
@@ -1,4 +1,4 @@
-// Generated by @wagmi/cli@0.1.10 on 2/26/2023 at 4:51:56 AM
+// Generated by @wagmi/cli@0.1.10 on 2/26/2023 at 5:21:08 AM
 import {
   useNetwork,
   useContract,
@@ -23,8 +23,8 @@ import {
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 /**
- * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
+ * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
+ * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
  * -
  */
 export const attestationStationABI = [
@@ -100,93 +100,24 @@ export const attestationStationABI = [
 ] as const
 
 /**
- * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
+ * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
+ * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
  * -
  */
 export const attestationStationAddress = {
-  10: '0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac',
-  420: '0x78991BB1D194C1235fe285240af8489CFA552151',
-  31337: '0xbe18A1B61ceaF59aEB6A9bC81AB4FB87D56Ba167',
+  10: '0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77',
+  420: '0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77',
+  31337: '0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77',
 } as const
 
 /**
- * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
+ * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
+ * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
  * -
  */
 export const attestationStationConfig = {
   address: attestationStationAddress,
   abi: attestationStationABI,
-} as const
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-// Counter
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/**
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Goerli Etherscan__](https://goerli.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
- * -
- */
-export const counterABI = [
-  {
-    type: 'event',
-    anonymous: false,
-    inputs: [
-      { name: 'from', internalType: 'address', type: 'address', indexed: true },
-      { name: 'to', internalType: 'address', type: 'address', indexed: true },
-      {
-        name: 'value',
-        internalType: 'uint256',
-        type: 'uint256',
-        indexed: false,
-      },
-    ],
-    name: 'Transfer',
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [],
-    name: 'increment',
-    outputs: [],
-  },
-  {
-    stateMutability: 'view',
-    type: 'function',
-    inputs: [],
-    name: 'number',
-    outputs: [{ name: '', internalType: 'uint256', type: 'uint256' }],
-  },
-  {
-    stateMutability: 'nonpayable',
-    type: 'function',
-    inputs: [{ name: 'newNumber', internalType: 'uint256', type: 'uint256' }],
-    name: 'setNumber',
-    outputs: [],
-  },
-] as const
-
-/**
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Goerli Etherscan__](https://goerli.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
- * -
- */
-export const counterAddress = {
-  1: '0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac',
-  5: '0x78991BB1D194C1235fe285240af8489CFA552151',
-  31337: '0xbe18A1B61ceaF59aEB6A9bC81AB4FB87D56Ba167',
-} as const
-
-/**
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Goerli Etherscan__](https://goerli.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
- * -
- */
-export const counterConfig = {
-  address: counterAddress,
-  abi: counterABI,
 } as const
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -219,8 +150,8 @@ export const semverABI = [
 /**
  * Wraps __{@link useContract}__ with `abi` set to __{@link attestationStationABI}__.
  *
- * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
+ * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
+ * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
  * -
  */
 export function useAttestationStation(
@@ -243,8 +174,8 @@ export function useAttestationStation(
 /**
  * Wraps __{@link useContractRead}__ with `abi` set to __{@link attestationStationABI}__.
  *
- * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
+ * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
+ * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
  * -
  */
 export function useAttestationStationRead<
@@ -279,8 +210,8 @@ export function useAttestationStationRead<
 /**
  * Wraps __{@link useContractRead}__ with `abi` set to __{@link attestationStationABI}__ and `functionName` set to `"attestations"`.
  *
- * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
+ * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
+ * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
  * -
  */
 export function useAttestationStationAttestations<
@@ -318,8 +249,8 @@ export function useAttestationStationAttestations<
 /**
  * Wraps __{@link useContractRead}__ with `abi` set to __{@link attestationStationABI}__ and `functionName` set to `"version"`.
  *
- * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
+ * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
+ * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
  * -
  */
 export function useAttestationStationVersion<
@@ -350,8 +281,8 @@ export function useAttestationStationVersion<
 /**
  * Wraps __{@link useContractWrite}__ with `abi` set to __{@link attestationStationABI}__.
  *
- * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
+ * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
+ * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
  * -
  */
 export function useAttestationStationWrite<
@@ -390,8 +321,8 @@ export function useAttestationStationWrite<
 /**
  * Wraps __{@link useContractWrite}__ with `abi` set to __{@link attestationStationABI}__ and `functionName` set to `"attest"`.
  *
- * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
+ * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
+ * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
  * -
  */
 export function useAttestationStationAttest<
@@ -434,8 +365,8 @@ export function useAttestationStationAttest<
 /**
  * Wraps __{@link usePrepareContractWrite}__ with `abi` set to __{@link attestationStationABI}__.
  *
- * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
+ * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
+ * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
  * -
  */
 export function usePrepareAttestationStationWrite<TFunctionName extends string>(
@@ -462,8 +393,8 @@ export function usePrepareAttestationStationWrite<TFunctionName extends string>(
 /**
  * Wraps __{@link usePrepareContractWrite}__ with `abi` set to __{@link attestationStationABI}__ and `functionName` set to `"attest"`.
  *
- * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
+ * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
+ * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
  * -
  */
 export function usePrepareAttestationStationAttest(
@@ -488,8 +419,8 @@ export function usePrepareAttestationStationAttest(
 /**
  * Wraps __{@link useContractEvent}__ with `abi` set to __{@link attestationStationABI}__.
  *
- * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
+ * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
+ * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
  * -
  */
 export function useAttestationStationEvent<TEventName extends string>(
@@ -513,8 +444,8 @@ export function useAttestationStationEvent<TEventName extends string>(
 /**
  * Wraps __{@link useContractEvent}__ with `abi` set to __{@link attestationStationABI}__ and `eventName` set to `"AttestationCreated"`.
  *
- * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
+ * - [__View Contract on Optimism Optimism Explorer__](https://explorer.optimism.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
+ * - [__View Contract on Optimism Goerli Etherscan__](https://goerli-optimism.etherscan.io/address/0xEE36eaaD94d1Cc1d0eccaDb55C38bFfB6Be06C77)
  * -
  */
 export function useAttestationStationAttestationCreatedEvent(
@@ -537,299 +468,6 @@ export function useAttestationStationAttestationCreatedEvent(
     typeof attestationStationABI,
     'AttestationCreated'
   >)
-}
-
-/**
- * Wraps __{@link useContract}__ with `abi` set to __{@link counterABI}__.
- *
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Goerli Etherscan__](https://goerli.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
- * -
- */
-export function useCounter(
-  config: Omit<UseContractConfig, 'abi' | 'address'> & {
-    chainId?: keyof typeof counterAddress
-  } = {} as any,
-) {
-  const { chain } = useNetwork()
-  const chainId = config.chainId ?? chain?.id
-  return useContract({
-    abi: counterABI,
-    address: counterAddress[chainId as keyof typeof counterAddress],
-    ...config,
-  })
-}
-
-/**
- * Wraps __{@link useContractRead}__ with `abi` set to __{@link counterABI}__.
- *
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Goerli Etherscan__](https://goerli.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
- * -
- */
-export function useCounterRead<
-  TFunctionName extends string,
-  TSelectData = ReadContractResult<typeof counterABI, TFunctionName>,
->(
-  config: Omit<
-    UseContractReadConfig<typeof counterABI, TFunctionName, TSelectData>,
-    'abi' | 'address'
-  > & { chainId?: keyof typeof counterAddress } = {} as any,
-) {
-  const { chain } = useNetwork()
-  const chainId = config.chainId ?? chain?.id
-  return useContractRead({
-    abi: counterABI,
-    address: counterAddress[chainId as keyof typeof counterAddress],
-    ...config,
-  } as UseContractReadConfig<typeof counterABI, TFunctionName, TSelectData>)
-}
-
-/**
- * Wraps __{@link useContractRead}__ with `abi` set to __{@link counterABI}__ and `functionName` set to `"number"`.
- *
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Goerli Etherscan__](https://goerli.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
- * -
- */
-export function useCounterNumber<
-  TSelectData = ReadContractResult<typeof counterABI, 'number'>,
->(
-  config: Omit<
-    UseContractReadConfig<typeof counterABI, 'number', TSelectData>,
-    'abi' | 'address' | 'functionName'
-  > & { chainId?: keyof typeof counterAddress } = {} as any,
-) {
-  const { chain } = useNetwork()
-  const chainId = config.chainId ?? chain?.id
-  return useContractRead({
-    abi: counterABI,
-    address: counterAddress[chainId as keyof typeof counterAddress],
-    functionName: 'number',
-    ...config,
-  } as UseContractReadConfig<typeof counterABI, 'number', TSelectData>)
-}
-
-/**
- * Wraps __{@link useContractWrite}__ with `abi` set to __{@link counterABI}__.
- *
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Goerli Etherscan__](https://goerli.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
- * -
- */
-export function useCounterWrite<
-  TMode extends WriteContractMode,
-  TFunctionName extends string,
-  TChainId extends number = keyof typeof counterAddress,
->(
-  config: TMode extends 'prepared'
-    ? UseContractWriteConfig<
-        TMode,
-        PrepareWriteContractResult<typeof counterABI, string>['abi'],
-        TFunctionName
-      > & { address?: `0x${string}`; chainId?: TChainId }
-    : UseContractWriteConfig<TMode, typeof counterABI, TFunctionName> & {
-        abi?: never
-        address?: never
-        chainId?: TChainId
-      } = {} as any,
-) {
-  const { chain } = useNetwork()
-  const chainId = config.chainId ?? chain?.id
-  return useContractWrite<TMode, typeof counterABI, TFunctionName>({
-    abi: counterABI,
-    address: counterAddress[chainId as keyof typeof counterAddress],
-    ...config,
-  } as any)
-}
-
-/**
- * Wraps __{@link useContractWrite}__ with `abi` set to __{@link counterABI}__ and `functionName` set to `"increment"`.
- *
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Goerli Etherscan__](https://goerli.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
- * -
- */
-export function useCounterIncrement<
-  TMode extends WriteContractMode,
-  TChainId extends number = keyof typeof counterAddress,
->(
-  config: TMode extends 'prepared'
-    ? UseContractWriteConfig<
-        TMode,
-        PrepareWriteContractResult<typeof counterABI, 'increment'>['abi'],
-        'increment'
-      > & {
-        address?: `0x${string}`
-        chainId?: TChainId
-        functionName?: 'increment'
-      }
-    : UseContractWriteConfig<TMode, typeof counterABI, 'increment'> & {
-        abi?: never
-        address?: never
-        chainId?: TChainId
-        functionName?: 'increment'
-      } = {} as any,
-) {
-  const { chain } = useNetwork()
-  const chainId = config.chainId ?? chain?.id
-  return useContractWrite<TMode, typeof counterABI, 'increment'>({
-    abi: counterABI,
-    address: counterAddress[chainId as keyof typeof counterAddress],
-    functionName: 'increment',
-    ...config,
-  } as any)
-}
-
-/**
- * Wraps __{@link useContractWrite}__ with `abi` set to __{@link counterABI}__ and `functionName` set to `"setNumber"`.
- *
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Goerli Etherscan__](https://goerli.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
- * -
- */
-export function useCounterSetNumber<
-  TMode extends WriteContractMode,
-  TChainId extends number = keyof typeof counterAddress,
->(
-  config: TMode extends 'prepared'
-    ? UseContractWriteConfig<
-        TMode,
-        PrepareWriteContractResult<typeof counterABI, 'setNumber'>['abi'],
-        'setNumber'
-      > & {
-        address?: `0x${string}`
-        chainId?: TChainId
-        functionName?: 'setNumber'
-      }
-    : UseContractWriteConfig<TMode, typeof counterABI, 'setNumber'> & {
-        abi?: never
-        address?: never
-        chainId?: TChainId
-        functionName?: 'setNumber'
-      } = {} as any,
-) {
-  const { chain } = useNetwork()
-  const chainId = config.chainId ?? chain?.id
-  return useContractWrite<TMode, typeof counterABI, 'setNumber'>({
-    abi: counterABI,
-    address: counterAddress[chainId as keyof typeof counterAddress],
-    functionName: 'setNumber',
-    ...config,
-  } as any)
-}
-
-/**
- * Wraps __{@link usePrepareContractWrite}__ with `abi` set to __{@link counterABI}__.
- *
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Goerli Etherscan__](https://goerli.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
- * -
- */
-export function usePrepareCounterWrite<TFunctionName extends string>(
-  config: Omit<
-    UsePrepareContractWriteConfig<typeof counterABI, TFunctionName>,
-    'abi' | 'address'
-  > & { chainId?: keyof typeof counterAddress } = {} as any,
-) {
-  const { chain } = useNetwork()
-  const chainId = config.chainId ?? chain?.id
-  return usePrepareContractWrite({
-    abi: counterABI,
-    address: counterAddress[chainId as keyof typeof counterAddress],
-    ...config,
-  } as UsePrepareContractWriteConfig<typeof counterABI, TFunctionName>)
-}
-
-/**
- * Wraps __{@link usePrepareContractWrite}__ with `abi` set to __{@link counterABI}__ and `functionName` set to `"increment"`.
- *
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Goerli Etherscan__](https://goerli.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
- * -
- */
-export function usePrepareCounterIncrement(
-  config: Omit<
-    UsePrepareContractWriteConfig<typeof counterABI, 'increment'>,
-    'abi' | 'address' | 'functionName'
-  > & { chainId?: keyof typeof counterAddress } = {} as any,
-) {
-  const { chain } = useNetwork()
-  const chainId = config.chainId ?? chain?.id
-  return usePrepareContractWrite({
-    abi: counterABI,
-    address: counterAddress[chainId as keyof typeof counterAddress],
-    functionName: 'increment',
-    ...config,
-  } as UsePrepareContractWriteConfig<typeof counterABI, 'increment'>)
-}
-
-/**
- * Wraps __{@link usePrepareContractWrite}__ with `abi` set to __{@link counterABI}__ and `functionName` set to `"setNumber"`.
- *
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Goerli Etherscan__](https://goerli.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
- * -
- */
-export function usePrepareCounterSetNumber(
-  config: Omit<
-    UsePrepareContractWriteConfig<typeof counterABI, 'setNumber'>,
-    'abi' | 'address' | 'functionName'
-  > & { chainId?: keyof typeof counterAddress } = {} as any,
-) {
-  const { chain } = useNetwork()
-  const chainId = config.chainId ?? chain?.id
-  return usePrepareContractWrite({
-    abi: counterABI,
-    address: counterAddress[chainId as keyof typeof counterAddress],
-    functionName: 'setNumber',
-    ...config,
-  } as UsePrepareContractWriteConfig<typeof counterABI, 'setNumber'>)
-}
-
-/**
- * Wraps __{@link useContractEvent}__ with `abi` set to __{@link counterABI}__.
- *
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Goerli Etherscan__](https://goerli.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
- * -
- */
-export function useCounterEvent<TEventName extends string>(
-  config: Omit<
-    UseContractEventConfig<typeof counterABI, TEventName>,
-    'abi' | 'address'
-  > & { chainId?: keyof typeof counterAddress } = {} as any,
-) {
-  const { chain } = useNetwork()
-  const chainId = config.chainId ?? chain?.id
-  return useContractEvent({
-    abi: counterABI,
-    address: counterAddress[chainId as keyof typeof counterAddress],
-    ...config,
-  } as UseContractEventConfig<typeof counterABI, TEventName>)
-}
-
-/**
- * Wraps __{@link useContractEvent}__ with `abi` set to __{@link counterABI}__ and `eventName` set to `"Transfer"`.
- *
- * - [__View Contract on Ethereum Etherscan__](https://etherscan.io/address/0x1A61839Eb5fC6eBBcAe01eD5E79062E598792Dac)
- * - [__View Contract on Goerli Etherscan__](https://goerli.etherscan.io/address/0x78991BB1D194C1235fe285240af8489CFA552151)
- * -
- */
-export function useCounterTransferEvent(
-  config: Omit<
-    UseContractEventConfig<typeof counterABI, 'Transfer'>,
-    'abi' | 'address' | 'eventName'
-  > & { chainId?: keyof typeof counterAddress } = {} as any,
-) {
-  const { chain } = useNetwork()
-  const chainId = config.chainId ?? chain?.id
-  return useContractEvent({
-    abi: counterABI,
-    address: counterAddress[chainId as keyof typeof counterAddress],
-    eventName: 'Transfer',
-    ...config,
-  } as UseContractEventConfig<typeof counterABI, 'Transfer'>)
 }
 
 /**
